+++ src/Components/Main/Pages/PublicationsPage/PublicationForm/PublicationForm.tsx
@@ -5,7 +5,7 @@ import { Button } from '../../../../Common/Button/Button';
 import { FormField } from '../../../../Common/FormField/FormField';
 import type { FormikErrorsType, FormikTouchedType, HandleChangeType } from '../../../../../utils/types/form';
 import { FormName } from '../../../../../utils/types/enums';
-import { submitFormOnKeyboardPress } from '../../../../../utils/helpers/componentHelpers';
+import { submitFormOnKeyboardPress } from '../../../../../utils/helpers/componentsHelpers';
 
 type PropsType = {
   handleChange: HandleChangeType;
